+++ zscaler/zid/services/users/users.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler"
-	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zidentity/services/common"
-	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zidentity/services/groups"
+	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zid/services/common"
+	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zid/services/groups"
 )
 
 const (